requirements.txt
#for parsing
beautifulsoup4==4.8.2
fake-useragent==0.1.11
PySocks==1.7.1
requests==2.23.0
regex==2020.4.4

#for DB
SQLAlchemy==1.3.15
alembic==1.4.2
psycopg2==2.8.4

#for telegram
python-telegram-bot[socks]==12.2.0

#for typing
yarl==1.4.2

#for env vars
python-dotenv==0.12.0

#for codestyle
black==19.10b0
isort==4.3.21
pycodestyle==2.5.0
